requirements.txt
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
django==5.0.7
django-cors-headers==4.4.0
google-generativeai==0.7.2
gunicorn==23.0.0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
whitenoise==6.7.0